rewritten brush engine

+++ src/main/java/pixelitor/tools/brushes/BrushShapeProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2014 Laszlo Balazs-Csiki
+ * Copyright 2015 Laszlo Balazs-Csiki
  *
  * This file is part of Pixelitor. Pixelitor is free software: you
  * can redistribute it and/or modify it under the terms of the GNU
@@ -8,11 +8,11 @@
  *
  * Pixelitor is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Pixelitor.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Pixelitor. If not, see <http://www.gnu.org/licenses/>.
  */
 package pixelitor.tools.brushes;
 
@@ -26,10 +26,10 @@ import java.awt.Shape;
 public enum BrushShapeProvider {
     HEART {
         @Override
-        Shape getShape(int x, int y, int width, int height) {
-            return new Heart(x, y, width, height);
+        Shape getShape(double x, double y, int width, int height) {
+            return new Heart((int) x, (int) y, width, height);
         }
     };
 
-    abstract Shape getShape(int x, int y, int width, int height);
+    abstract Shape getShape(double x, double y, int width, int height);
 }
